+++ frontend/lib/app.dart
@@ -1,7 +1,7 @@
 
 
 import 'package:flutter/material.dart';
-import 'package:frontend/domain/utils/theme.dart';
+import 'package:frontend/presentation/utils/theme.dart';
 import 'package:frontend/presentation/assets/l10n/generated/l10n.dart';
 import 'package:frontend/presentation/screens/board/on_board.dart';
 import 'package:flutter_localizations/flutter_localizations.dart';
@@ -30,7 +30,7 @@ class _AppState extends State<App> {
     return  MaterialApp(
         debugShowCheckedModeBanner: false,
         theme: AppTheme.lightTheme, // Apply the light theme
-        darkTheme: AppTheme.darkTheme, // Apply the dark theme (optional)
+        darkTheme: AppTheme.darkTheme,
         title: 'HeartBeats',
          locale: _locale,
       supportedLocales: S.delegate.supportedLocales,